+++ CHANGELOG.md
@@ -7,10 +7,20 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+### Changed
+### Fixed
+
+## [0.3.0] - 2019-06-23
+### Added
 - New behaviours: `keyup` and `lose_focus`.
+- Select changes to make use of the `Choices` javascript library instead of the `Selectr` library.
+- Select: `min_charwidth` option to set the `min-width` style in rem.
+- Select: `native` option to render as a native select tag.
+- Select: `remove_search_for_small_list` option to automatically render without a search box if the list of options is small.
+- Select: `sort_items` option to have the javascript control sort items.
+- Select: if `prompt` is not set, ensure the dropdown is not clearable.
 ### Changed
 - Label renders as a div instead of a disabled input. This allows selecting of the text in the label.
-### Fixed
 
 ## [0.2.2] - 2019-06-03
 ### Added